Module 1-Tezos_protocol_environment.S

module type T = sig ... end

Generic interface for a datatype with comparison, pretty-printer and serialization functions.

module type HASHABLE = sig ... end

Generic interface for a datatype with comparison, pretty-printer, serialization functions and a hashing function.

Hash Types

The signature of an abstract hash type, as produced by functor Make_SHA256. The t type is abstracted for separating the various kinds of hashes in the system at typing time. Each type is equipped with functions to use it as is of as keys in the database or in memory sets and maps.

module type MINIMAL_HASH = sig ... end

module type RAW_DATA = sig ... end

module type B58_DATA = sig ... end

module type ENCODER = sig ... end

module type SET = sig ... end

module type MAP = sig ... end

module type INDEXES = sig ... end

module type HASH = sig ... end

module type MERKLE_TREE = sig ... end

module type SIGNATURE = sig ... end